Data loader & MNIST examples & various perf improvements (#336)

+++ alpa/pipeline_parallel/decentralized_distributed_runtime.py
@@ -842,11 +842,8 @@ class DecentralizedDistributedRuntime(BaseDistributedRuntime):
                         self.mesh_index_to_outvar_indices_mapping[mesh_idx]
                         [i]].aval.dtype
                     output_bufs[mesh_idx][i][j] = RemoteBufferRef(
-                        physical_mesh,
-                        host_id,
-                        device_id,
-                        output_uuid_transposed[i][host_id][device_id],
-                        dtype=dtype)
+                        physical_mesh, host_id, device_id,
+                        output_uuid_transposed[i][host_id][device_id])
         return self.outs_handler(output_bufs)
 
     def _setup_outs_handler(self):
